andy's changes

+++ 09-Capstone/student/java/tenmo-client/src/main/java/com/techelevator/tenmo/models/Transfer.java
@@ -8,11 +8,11 @@ public class Transfer {
 	private int transfer_id;
 	private int transfer_type_id;
 	private int transfer_status_id;
-	private BigDecimal amount;
+	private int amount;
 	private int account_to;
 	private int account_from;
 	
-	public Transfer(int account_from, int account_to, BigDecimal amount) {
+	public Transfer(int account_from, int account_to, int amount) {
 		this.amount = amount;
 		this.account_to = account_to;
 		this.account_from = account_from;
@@ -26,10 +26,10 @@ public class Transfer {
 		this.account_from = account_from;
 	}
 
-	public BigDecimal getAmount() {
+	public int getAmount() {
 		return amount;
 	}
-	public void setAmount(BigDecimal amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
 	public int getAccount_to() {
